pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "ballast"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "ag-ui-protocol",
    "ag-ui-langgraph",
    "langgraph",
    "langchain-openai",
    "langchain-anthropic",
    "anthropic>=0.20",
    "httpx>=0.27",
    "filelock",
    "opentelemetry-api>=1.30",
    "pydantic>=2.0",
    "pydantic-ai>=0.0.13,<1.0",
    "python-dotenv",
    "fastapi>=0.100",
    "uvicorn>=0.27",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-asyncio",
    "textual>=8.0",
]

[tool.hatch.build.targets.wheel]
packages = ["ballast"]

[tool.pytest.ini_options]
markers = [
    "integration: requires ANTHROPIC_API_KEY and live Anthropic API access — skip with -m 'not integration'",
]
